architecture rework for future updates/features
