+++ .gitignore
@@ -1,0 +1,7 @@
+.vscode/
+node_modules/
+
+.env
+*.bak
+
+
