+++ extension/popup/index.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+   <meta charset="UTF-8">
+   <meta name="viewport" content="width=device-width, initial-scale=1.0">
+   <title>Popup</title>
+</head>
+<body>
+   <div class="container">
+      <h1>Popup Title</h1>
+   
+   </div> 
+   <script src="index.js"></script>
+</body>
+</html>
