+++ package.json
@@ -1,0 +1,16 @@
+{
+  "name": "chrome-vxtwitter",
+  "version": "1.0.0",
+  "description": "<p align=\"center\">     <h1 align=\"center\">🤖 AutoTwitFix 🤖</h1> </p> <p align=\"center\">for Twitter</p> I stole the majority of the code from this <a href=\"https://addons.mozilla.org/fr/firefox/addon/vx-twitter/\">firefox add-on</a>. <p>   <h2>Features</h2>     <ul>       <li>Upon clicking \"copy the link\" on the \"share\" context menu of a video post, the link is automatically converted to a \"vxtwitter\" version in your clipboard.</li>       <li>When you are already on a tweet's page, you can use CTRL + C and it copies the vxtwitter link automatically.</li>     </ul> </p>",
+  "main": "main.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "devDependencies": {
+    "webpack": "^5.88.1",
+    "webpack-cli": "^5.1.4"
+  }
+}
