+++ src/utils/clipboard.js
@@ -1,0 +1,9 @@
+async function writeToClipboard(text) {
+   try {
+      await navigator.clipboard.writeText(text)
+   } catch (e) {
+      // ignore error
+   }
+}
+
+export { writeToClipboard };
